+++ CHANGELOG.md
@@ -3,7 +3,8 @@ Changes to the schema are documented in this file.
 
 ## Current (updates before releasing)
 ### Added
-  - N/A
+  - `relevant protocols` and `funding sources` slots
+  - add amplicon slots to omics processing (#139)
 ### Fixed
   - N/A
 ### Changed